pyproject.toml
[tool.poetry]
name = "peakrdl_bsv"
version = "1.0.0"
description = "Export BSV description from the systemrdl-compiler register model"
authors = ["vijayvithal jahagirdar <jahagirdar.vs@gmail.com>"]
maintainers = ["vijayvithal jahagirdar <jahagirdar.vs@gmail.com>"]
license = "BSD2-only"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
]
repository = "https://github.com/jahagirdar/peakrdl-bsv.git"
documentation = "https://peakrdl-bsv.readthedocs.io"
readme = "README.md"
packages = [{ include = "peakrdl_bsv", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10.0"
systemrdl-compiler = "^1.25.0"
peakrdl = "^1.1.0"
py-markdown-table = "^0.3.3"

# Dependabot force updates for Python >3.6
gitpython = { version = "^3.1.41", python = ">=3.7" }
setuptools = { version = "^65.5.1", python = ">=3.7" }
pygments = { version = "^2.15.0", python = ">=3.7" }
jinja2 = { version = "^3.1.3", python = ">=3.7" }

[tool.poetry.group.dev.dependencies]
pre-commit = [
    { version = "^2.0.0", python = ">=3.7,<3.9" },
    { version = "^3.7.0", python = ">=3.9" },
]

[tool.poetry.group.lint.dependencies]
black = { version = "^24.4.0", python = ">=3.8" }
packaging = { version = "^24.0", python = ">=3.8" } # Required for black dependency resolution.
isort = { version = "^5.11.0", python = ">=3.7" }
mypy = "^0.971"
pydocstyle = "^6.1.1"
ruff = { version = "^0.3.7", python = ">=3.7" }

[tool.poetry.group.test.dependencies]
pytest = [
    { version = "^7.0.0", python = "<3.7" },
    { version = "^7.2.0", python = ">3.7,<=3.8" },
    { version = "^8.1.1", python = ">3.8" },
]
pytest-cov = [
    { version = "^4.0.0", python = "<3.8" },
    { version = "^5.0.0", python = ">=3.8" },
]
coverage = { version = "^6.3", python = ">=3.7" }
coveralls = "^3.3.1"

# Dependabot force updates for Python >3.6
requests = { version = "^2.31.0", python = ">=3.7" }

[tool.poetry.group.doc.dependencies]
linuxdoc = { version = "^20211220", python = ">=3.9" }
Sphinx = { version = "^7.2.6", python = ">=3.9" }
sphinx-rtd-theme = { version = "^2.0.0", python = ">=3.9" }
m2r2 = { version = "^0.3.3", python = ">=3.9" }
pygments-systemrdl = { version = "^1.2.0", python = ">=3.9"}

[tool.poetry.plugins."peakrdl.exporters"]
"bsv" = "peakrdl_bsv.__peakrdl__:Exporter"


[tool.pdm]
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pydocstyle]
convention = "pep257"

[project]
authors = [
    {name = "vijayvithal jahagirdar", email = "jahagirdar.vs@gmail.com"},
]
maintainers = [
    {name = "vijayvithal jahagirdar", email = "jahagirdar.vs@gmail.com"},
]
license = {text = "BSD2-only"}
requires-python = "<4.0.0,>=3.8.0"
dependencies = [
    "systemrdl-compiler<2.0.0,>=1.25.0",
    "peakrdl<2.0.0,>=1.1.0",
    "py-markdown-table<1.0.0,>=0.3.3",
    "gitpython<4.0.0,>=3.1.41; python_version >= \"3.7\"",
    "setuptools<66.0.0,>=65.5.1; python_version >= \"3.7\"",
    "pygments<3.0.0,>=2.15.0; python_version >= \"3.7\"",
    "jinja2<4.0.0,>=3.1.3; python_version >= \"3.7\"",
    "Markdown>=3.7",
]
name = "peakrdl_bsv"
version = "1.0.0"
description = "Export BSV description from the systemrdl-compiler register model"
classifiers = ["Development Status :: 5 - Production/Stable", "Intended Audience :: Developers", "Operating System :: OS Independent", "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)"]
readme = "README.md"

[project.urls]
repository = "https://github.com/jahagirdar/peakrdl-bsv.git"
documentation = "https://peakrdl-bsv.readthedocs.io"

[project.entry-points."peakrdl.exporters"]
bsv = "peakrdl_bsv.__peakrdl__:Exporter"
